Update pages with Map component
