+++ client/src/pages/Home.js
@@ -1,23 +1,25 @@
 import React from 'react';
+import "./styles.css";
 import NavbarComponent from '../components/Navbar'
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Carousel from 'react-bootstrap/Carousel';
 import Container from 'react-bootstrap/Container';
-
+import Map from '../components/Map'
 
 export default function Home() {
     return (
         <div>
         <NavbarComponent/>
+        <div className='top-height'>
         <Container align="center">
             <Container>
-                <h1>Welcome to VanDons!</h1>
+                <h1>Welcome to VanDonates!</h1>
             </Container>
             <Container>
                 <Row>
                     <Col>
-                        Pane 1
+                        <p>At VanDonates you </p>
                     </Col>
                     <Col>
                     <Carousel variant="dark">
@@ -61,6 +63,8 @@ export default function Home() {
                 </Row>
             </Container>
         </Container>
+        </div>
+        <Map/>
         </div>
     )
 }
