+++ client/src/pages/styles.css
@@ -1,3 +1,12 @@
-.top-list{
-    height: 40vh;
+.top-height{
+    height: 43vh;
+}
+
+p{
+    font-size: 1.2rem;
+}
+
+h1{
+    margin-top: 100px;
+    margin-bottom: 10px;
 }
